+++ wkz/activity_views.py
@@ -17,8 +17,8 @@ from wkz.awards_views import (
 )
 from wkz.best_sections.generic import activity_suitable_for_awards
 from wkz.forms import DATETIMEPICKER_FORMAT, AddActivityForm, EditActivityForm
+from wkz.io.gpx_exporter import save_activity_to_gpx_file
 from wkz.models import Activity, BestSection, Lap, Sport
-from wkz.parser.gpx_exporter import save_activity_to_gpx_file
 from wkz.plotting.plot_time_series import plot_time_series
 from wkz.tools.style import Style
 from wkz.views import MapView, get_all_form_field_ids
