+++ wkz/io/auto_naming.py
